requirements.txt
Flask>=2.0,<3.0
numpy>=1.20,<2.0
opencv-python>=4.5,<5.0
opencv-python-headless>=4.5,<5.0
Pillow>=8.3,<10.0